+++ sphinx/source/changelog.rst
@@ -8,6 +8,8 @@ Changelog
   tape block (which are usually, but not always, the flag and parity bytes)
 * Added support to :ref:`tap2sna.py` for modifying memory with XOR and ADD
   operations
+* Added the ``--clear`` option to :ref:`bin2tap.py` (to use a CLEAR command in
+  the BASIC loader and leave the stack pointer alone)
 * Added support to :ref:`skool2ctl.py` and :ref:`skool2sft.py` for preserving
   characters in DEFW statements (e.g. ``DEFW "!"``)
 * Added support for characters in DEFS statements (e.g. ``DEFS 10,"!"``)